% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{assaggi2020}
\alias{assaggi2020}
\title{Panel test degli oli}
\format{
Un dataset con 27 righe e 11 variabili:
\describe{
  \item{Codice_azienda}{codice identificativo dell'azienda (es. SA_01)}
  \item{Anno}{l'anno del campionamento (es. 2020)}
  \item{Tipo_olio}{informazione addizionale sul tipo di olio. Alcune aziende hanno fornito diversi oli a diverse fasi di lavorazione.}
  \item{Categoria_olio}{categoria di appartenenza dell'olio (lampante, vergine o extravergine)}
  \item{Mediana_fruttato}{mediana del fruttato dell'olio (1.0 - 6.2)}
  \item{CVR_Fruttato}{il coefficiente di variazione robusta del fruttato. Se il valore è inferiore al limite di legge è riportato "ok", altrimenti "nok"}
  \item{Presenza_difetti}{presenza di difetti nell'olio (no, si)}
  \item{Mediana_amaro}{mediana dell'amaro dell'olio (1.9 - 4.5)}
  \item{CVR_amaro}{il coefficiente di variazione robusta dell'amaro. Se il valore è inferiore al limite di legge è riportato "ok", altrimenti "nok"}
  \item{Mediana_piccante}{mediana dell'amaro dell'olio (1.5 - 4.8)}
  \item{CVR_piccante}{il coefficiente di variazione piccante dell'amaro. Se il valore è inferiore al limite di legge è riportato "ok", altrimenti "nok"}
  
}
}
\source{
\url{http://bioinfo.na.iac.cnr.it/olivehealth/File_csv.rar}
}
\usage{
assaggi2020
}
\description{
Un dataset che contiene i risultati del panel test effettuato sugli oli. Non tutte le aziende hanno partecipato.
}
\keyword{datasets}
