% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drupecamp2020}
\alias{drupecamp2020}
\title{Schede del campionamento 2020 delle drupe}
\format{
Un dataset con 52 righe e 6 variabili:
\describe{
  \item{Codice_azienda}{codice identificativo dell'azienda (es. SA_01)}
  \item{N_campionamento}{codice utilizzato per distinguere il primo dal secondo campionamento (R1, R2)}
  \item{Data_campionamento}{la data del campionamento nel formato yyyy-mm-gg (nel file csv è gg/mm/yyyy)}
  \item{Fase_fenologica}{fase fenologica della pianta espresso come numero-factor (51, 55, 59, 61, 65, 69, 71, 75, 79, 81, 85, 89)}
  \item{Indice_maturazione}{il livello di maturazione delle drupe espresso come numero-factor (0, 1, 2, 3, 4, 5, 6, 7, 8)}
  \item{Note}{eventuali note riguardanti il campionamento}
}
}
\source{
\url{http://bioinfo.na.iac.cnr.it/olivehealth/File_csv.rar}
}
\usage{
drupecamp2020
}
\description{
Un dataset che contiene le schede del campionamento 2020 delle drupe. Le variabili sono le seguenti:
}
\keyword{datasets}
