% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_heatmap.R
\name{make_heatmap}
\alias{make_heatmap}
\title{Make a Heatmap using ComplexHeatmap package.}
\usage{
make_heatmap(
  datasorted,
  add_annot,
  scale_data,
  row_dend,
  row_nclust,
  col_dend,
  col_nclust,
  dist_method,
  clust_method
)
}
\arguments{
\item{datasorted}{A dataframe}

\item{add_annot}{A column used for the annotation (e.g. "Provincia")}

\item{scale_data}{Scaling data option. Data can be scaled by row ("row"), by column ("column") or not scaled ("none").}

\item{row_dend}{Option for dendogram by row (TRUE or FALSE).}

\item{row_nclust}{Number of cluster in the row dendrogram (a number).}

\item{col_dend}{Option for dendogram by column (TRUE or FALSE).}

\item{col_nclust}{Number of cluster in the column dendrogram (a number).}

\item{dist_method}{Choose a distance method (e.g "euclidean"). See stats::dist() help.}

\item{clust_method}{Choose a clustering method (e.g "ward.D"). See stats::hclust() help.}
}
\description{
This function render a Heatmap using ComplexHeatmap package.
}
\examples{
\dontrun{

make_heatmap(
datasorted = dtheatsorted(), 
add_annot = input$selectannot, 
scale_data = input$selscaleheat, 
dist_method = input$seldistheatpol, 
clust_method = input$selhclustheatpol,
row_dend = input$rowdend, 
row_nclust = input$sliderrowheat,
col_dend = input$columndend, 
col_nclust = input$slidercolheat
)


}
}
