% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tmap.R
\name{make_tmap}
\alias{make_tmap}
\title{Make a tmap object that can be rendered in shiny}
\usage{
make_tmap(pointcoord, datainfo, dotlegend, shp)
}
\arguments{
\item{pointcoord}{A dataset filled with two columns (UTM_33T_N and UTM_33T_E). Each row is a point with its UTM 33T cordinates.}

\item{datainfo}{A dataset with points informations. These informations will be shown on the map.}

\item{dotlegend}{An informative column used for the legend and for the mapping of the dots (e.g. "Polifenoli_totali" or "Azienda").}

\item{shp}{A shapefile used for the background in the UTM 33T coordinate system.}
}
\description{
This function render a tmap using a shapefile (campania) as background and a table with coordinates for the points.
}
\examples{
\dontrun{

output$map1 = make_tmap(
 pointcoord = filtereddata(), 
 datainfo = datmap1(), 
 dotlegend = showcolumn2(), 
 shp = campania)
}



}
