% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Olv_select_col.R
\name{Olv_select_col}
\alias{Olv_select_col}
\title{Select column from a dataframe}
\usage{
Olv_select_col(data, input)
}
\arguments{
\item{data}{The dataset used to select the column. Should be a reactive value (e.g. datmap1())}

\item{input}{the input from the ui (e.g. the input from a \code{selectInput()}) in the form of \code{input$...}.}
}
\description{
Pick a column from a dataframe using an input from the UI. Can be for example from a \code{selectInput()}.
}
\examples{
\dontrun{

  showcolumn2 = reactive({
  
  Olv_select_col(
  data = datmap1(),       #a dataset made with read_csv() 
  input = input$select3   #the input from a shiny::selectInput()
  )
  
  })

}

}
