% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poliftot2020}
\alias{poliftot2020}
\title{Polifenoli totali di drupe, foglie, olio, posa e sansa 2020}
\format{
Una lista con 4 dataset:
\itemize{
\item{\strong{Foglie}}
\describe{
Un dataset con 312 righe e 5 variabili:
  \item{Codice_azienda}{codice identificativo dell'azienda (es. SA_01)}
  \item{N_campionamento}{codice utilizzato per distinguere il primo dal secondo campionamento (R1, R2)}
  \item{Anno}{l'anno del campionamento (es. 2020)}
  \item{Estrazione}{codice identificativo dell'estrazione}
  \item{`Polifenoli (mg/g foglie)`}{polifenoli totali delle foglie in mg su g di foglie}
  }
 \item{\strong{Drupe}}
 \describe{
 Un dataset con 312 righe e 6 variabili:
  \item{Codice_azienda}{codice identificativo dell'azienda (es. SA_01)}
  \item{N_campionamento}{codice utilizzato per distinguere il primo dal secondo campionamento (R1, R2)}
  \item{Anno}{l'anno del campionamento (es. 2020)}
  \item{Estrazione}{codice identificativo dell'estrazione}
  \item{Presenza_larve}{presenza di larve (0 niente larve; 1 poche larve; 2 molte larve)}
  \item{`Polifenoli (mg/g drupe)`}{polifenoli totali delle drupe in mg su g di drupe}
 }
 \item{\strong{Olio}}
 \describe{
  \item{Codice_azienda}{codice identificativo dell'azienda (es. SA_01)}
  \item{N_campionamento}{codice utilizzato per distinguere il primo dal secondo campionamento (R1, R2)}
  \item{Tipo_olio}{colonna che identifica se le drupe sono state denocciolate ("denocciolato") o no ("massa_ultima")}
  \item{Anno}{l'anno del campionamento (es. 2020)}
  \item{Estrazione}{codice identificativo dell'estrazione}
  \item{`Polifenoli (mg/g olio)`}{polifenoli totali delle olio in mg su kg di olio}
 }
 \item{\strong{Posa}}
 \describe{
  \item{Codice_azienda}{codice identificativo dell'azienda (es. SA_01)}
  \item{N_campionamento}{codice utilizzato per distinguere il primo dal secondo campionamento (R1, R2)}
  \item{Tipo_olio}{colonna che identifica se le drupe sono state denocciolate ("denocciolato") o no ("massa_ultima")}
  \item{Anno}{l'anno del campionamento (es. 2020)}
  \item{Estrazione}{codice identificativo dell'estrazione}
  \item{`Polifenoli (mg/g posa)`}{polifenoli totali della posa in mg su kg di posa}
 }
 \item{\strong{Sansa}}
 \describe{
  \item{Codice_azienda}{codice identificativo dell'azienda (es. SA_01)}
  \item{N_campionamento}{codice utilizzato per distinguere il primo dal secondo campionamento (R1, R2)}
  \item{Tipo_olio}{colonna che identifica se le drupe sono state denocciolate ("denocciolato") o no ("massa_ultima")}
  \item{Anno}{l'anno del campionamento (es. 2020)}
  \item{Estrazione}{codice identificativo dell'estrazione}
  \item{`Polifenoli (mg/g sansa)`}{polifenoli totali della sansa in mg su kg di sansa}
 }
}
}
\source{
\url{http://bioinfo.na.iac.cnr.it/olivehealth/File_csv.rar}
}
\usage{
poliftot2020
}
\description{
Una lista che contiene 4 dataframe contenenti i dati sui polifenoli totali  di drupe, foglie, olio, posa e sansa della stagione 2020.
}
\keyword{datasets}
