% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_OliveHealthR.R
\name{run_OliveHealthR}
\alias{run_OliveHealthR}
\title{Run the Shiny Application}
\usage{
run_OliveHealthR(...)
}
\arguments{
\item{...}{A series of options to be used inside the app.}
}
\description{
Run the Shiny Application
}
