% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_photo.R
\name{download_photo}
\alias{download_photo}
\title{Download photos and copy in www folder}
\usage{
download_photo()
}
\arguments{
\item{url}{Url where download the zip file. By default "https://drive.google.com/u/0/uc?export=download&confirm=_w5z&id=1cLQn3Xhs5aH0hVBrzRksBC0m5Nvk7Zpk".}
}
\description{
This function download photos from a link and copy them in the www folder. Data are uploaded on https://gofile.io/uploadFiles.
}
