% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_photo.R
\name{download_photo}
\alias{download_photo}
\title{Download photos and copy in www folder}
\usage{
download_photo()
}
\arguments{
\item{url}{Url where download the zip file. The default is "http://bioinfo.na.iac.cnr.it/olivehealth/foto_drupe_foglie_20_21.zip".}
}
\description{
This function download photos from a link and copy them in the www folder. Data are uploaded on a CNR server.
}
