% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_photo.R
\name{download_photo}
\alias{download_photo}
\title{Download photos and copy in www folder}
\usage{
download_photo()
}
\arguments{
\item{url}{Url where download the zip file. By default "https://srv-store4.gofile.io/download/Vhtmor/9e840835a48fcd254263fdc03dbf690d".}
}
\description{
This function download photos from a link and copy them in the www folder. Data are uploaded on https://gofile.io/uploadFiles.
}
