% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_heatmap.R
\name{make_heatmap}
\alias{make_heatmap}
\title{Make a Heatmap using ComplexHeatmap package.}
\usage{
make_heatmap(
  datasorted,
  add_annot,
  scale_data,
  row_dend,
  row_nclust,
  col_dend,
  col_nclust,
  dist_method,
  clust_method,
  col_lab,
  unit_legend,
  year_presence = TRUE,
  col_label_size = 13.2,
  bordi = c(2, 2, 2, 15)
)
}
\arguments{
\item{datasorted}{A dataframe done with sorder_data() function. See in ?sorder_data() the "data2" parameter for help.}

\item{add_annot}{A column used for the annotation (e.g. "Provincia")}

\item{scale_data}{Scaling data option. Data can be scaled (using scale()) by row ("row"), by column ("column") or not scaled ("none"). The result is the z-score.}

\item{row_dend}{Option for dendogram by row (TRUE or FALSE).}

\item{row_nclust}{Number of cluster in the row dendrogram (a number).}

\item{col_dend}{Option for dendogram by column (TRUE or FALSE).}

\item{col_nclust}{Number of cluster in the column dendrogram (a number).}

\item{dist_method}{Choose a distance method (e.g "euclidean"). See stats::dist() help.}

\item{clust_method}{Choose a clustering method (e.g "ward.D"). See stats::hclust() help.}

\item{col_lab}{A label for the column (i.e. "Polifenoli" or "Misure").}

\item{unit_legend}{An unit label for the legend. If data are scaled will be "Z-score" otherwise will be the input.}

\item{year_presence}{Option for data without a year column. Default is TRUE (year column present).}

\item{col_label_size}{Size of the column labels. By default col_label_size = 13.2.}
}
\description{
This function render a Heatmap using ComplexHeatmap package.
}
\examples{
\dontrun{

make_heatmap(
datasorted = dtheatsorted(), 
add_annot = input$selectannot, 
scale_data = input$selscaleheat, 
dist_method = input$seldistheatpol, 
clust_method = input$selhclustheatpol,
row_dend = input$rowdend, 
row_nclust = input$sliderrowheat,
col_dend = input$columndend, 
col_nclust = input$slidercolheat
)


}
}
