% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_meteo.R
\name{makedata_meteo}
\alias{makedata_meteo}
\title{Make a tmap object that can be rendered in shiny}
\usage{
makedata_meteo(ncfile, dati_aziende)
}
\arguments{
\item{ncfile}{Path of the ncfile. "C:\\Users\\fabio\\Desktop\\OliveHealthR\\data-raw\\precipitazioni\\precipit_e_altro_20_21.nc"}

\item{data}{A dataframe with the coordinates columns(UTM_33T_N and UTM_33T_E) and relative infos. Use data()}
}
\description{
This function render a tmap using a shapefile (campania) as background and a table with coordinates for the points.
}
\examples{
\dontrun{

output$map1 = make_tmap(
 data = data = dtdrupfilt(), dotlegend = showcolumnmap2()
 )
}



}
