% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_meteo.R
\name{makedata_meteo}
\alias{makedata_meteo}
\title{Legge netcdf da ERA5-LAND mensile e assegna i valori ai punti delle aziende}
\usage{
makedata_meteo(ncfile, dati_aziende)
}
\arguments{
\item{ncfile}{ncfile path}

\item{dati_aziende}{A dataframe with the coordinates columns(UTM_33T_N and UTM_33T_E) and relative infos. Use data()}
}
\description{
Questa funzione Legge netcdf da ERA5-LAND mensile e assegna i valori ai punti delle aziende
}
