% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{morfometria2020}
\alias{morfometria2020}
\title{Morfometria di drupe, foglie ed endocarpo della stagione 2020}
\format{

}
\source{
\url{http://bioinfo.na.iac.cnr.it/olivehealth/File_csv.rar}
}
\usage{
morfometria2020
}
\description{
Una lista che contiene 4 vettori contenenti i dati morfometrici di foglie, endocarpo, drupe e rapporti drupe-endocarpo della stagione 2020.
}
\keyword{datasets}
