% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_OliveHealthR.R
\name{run_OliveHealthR}
\alias{run_OliveHealthR}
\title{Launches the web shiny App "OliveHealthR"}
\usage{
run_OliveHealthR()
}
\description{
Simple function to connect UI and Server of the shiny App and launch it inside the default browser.
}
\note{
OliveHealthR is builded with {golem}.
}
\examples{
\dontrun{
libraray(OliveHealthR)
run_OliveHealthR()
}

}
